Fix crash: pass Float to percent_format string resource
TorrentDetailsFragment's percent_format calls were passing an Int
(progressPercent) into a %.1f format string, throwing
IllegalFormatConversionException as soon as the torrent details
screen received its first progress update (e.g. right after adding
a magnet). Pass progressPercent.toFloat() instead, matching the
pattern already used correctly in TorrentListPagingAdapter.

+++ app/app/src/main/java/com/github/livingwithhippos/unchained/torrentdetails/view/TorrentDetailsFragment.kt
@@ -124,8 +124,9 @@ class TorrentDetailsFragment : UnchainedFragment(), TorrentContentListener {
                     binding.tvName.text = torrent.name
                     val progressPercent = (torrent.progress * 100).toInt()
                     binding.tvProgressPercent.text =
-                        getString(R.string.percent_format, progressPercent)
-                    binding.tvProgress.text = getString(R.string.percent_format, progressPercent)
+                        getString(R.string.percent_format, progressPercent.toFloat())
+                    binding.tvProgress.text =
+                        getString(R.string.percent_format, progressPercent.toFloat())
                     if (progressPercent in 0..99) {
                         binding.tvProgress.visibility = View.VISIBLE
                     } else {
